+++ src/chat-images.ts
@@ -1,9 +1,8 @@
 'use strict';
 
 import {log, MODULE_NAME} from "./module/utils";
-import {createPopoutOnClick} from "./module/image-manager/popout-manager";
-import {convertMessageToImage, handleChatInteraction} from "./module/image-manager/image-manager";
 import {createUploadFolderIfMissing, getSetting, registerSettings, UPLOAD_FOLDER_PATH} from "./module/settings";
+import {convertMessageToImage, handleChatInteraction, createPopoutOnClick} from "./module/image-manager/manager";
 
 Hooks.once('init', () => {
     // register all the module's settings
