+++ src/module/image-manager/image-extraction.ts
